% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpattern.R
\name{kpattern}
\alias{kpattern}
\title{Kpattern}
\usage{
kpattern(shapefile, longitude = longitude, latitude = latitude, nsim = 99)
}
\description{
Utilizes the Kest and Envelope functions from Spatstat and creates a visualization with
ggplot2.
}
\details{
kpattern = function(shapefile, data, nsim = 99)
}
