% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpattern.R
\name{mpattern}
\alias{mpattern}
\title{mpattern}
\usage{
mpattern(
  shapefile,
  longitude = longitude,
  latitude = latitude,
  title = NULL,
  pcolor = "black",
  psize = 0.95
)
}
\description{
Creates a visualization of point pattern data over a shapefile.
}
