% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg.R
\name{plotgg}
\alias{plotgg}
\title{ggplot2 map syntax generator}
\usage{
plotgg(
  shapefile = shapefile,
  adata = adata,
  maptitle = "Map Title",
  continuous = TRUE,
  guide.title = "Guide Title"
)
}
\description{
ggplot2 map syntax generator
}
