% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_qspatial.R
\name{theme_qspatial}
\alias{theme_qspatial}
\title{A ggplot2 theme for qspatial!}
\usage{
theme_qspatial(base_size = 11, base_family = "Arial",
  base_line_size = base_size/170, base_rect_size = base_size/170)
}
\description{
A ggplot2 theme for qspatial!
}
