% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areamap.R
\name{areamap}
\alias{areamap}
\title{Area map function}
\usage{
areamap(shapefile = shapefile, adata = data, maptitle = "Map Title",
  guidetitle = "Guide Title", continuous = TRUE)
}
\description{
Area map function
}
