% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmpattern.R
\name{qmpattern}
\alias{qmpattern}
\title{Quick mapping for point pattern data}
\usage{
qmpattern(shapefile, longitude = longitude, latitude = latitude,
  fun = c("G", "F"), sigma = 0.01, nsim = 99, palette = "RdYlBu",
  psize = 0.75)
}
\arguments{
\item{shapefile}{A shapefile of the study region.}

\item{longitude}{The name of the variable containing the longitude.}

\item{latitude}{The name of the variable containing the latitude.}

\item{fun}{A vector with two summary functions. "G" for nearest neighbor function, "F" for empty space function, "K" for Ripley's K function and "J" for the J function.}

\item{sigma}{Optional. The radius used to estimate the Intensity, if a radius is not chosen it will be calculated using the bw.diggle function.}

\item{nsim}{The number of simulations generated when computing the envelope. The default is 99.}

\item{palette}{The color pallete used in the Intensity plot.}
}
\description{
Generates the visualization and analysis of first and second order properties of point pattern data.
}
\details{
The function generates a quick visualization and analysis of point pattern data with
ggplot2 and spatstat's summary functions. The default option is to analyze using
the G and F functions, they work well for larger datasets and the simulation does not
take much time to complete. There are options included to pick two functions between
K, G, F and J summary functions, but it's important to notice that the runtime
of the K function is considerably bigger.
}
